Updated button component to make the entire button clickable
Previously, only the text of the button was clickable in the button component. To improve user experience and make it more intuitive, I made changes to the button component so that the entire button is now clickable. This change allows users to click anywhere within the button element to trigger the desired action.